refactor(runtime): extract scaffold module from plugin_manager
The plugin-scaffolding cluster moves into `plugin_manager/scaffold.rs`
— `scaffold_plugin` (pub re-exported from the parent so
`librefang_runtime::plugin_manager::scaffold_plugin` still resolves
for the kernel callers), `HookFiles`, `hook_templates`, and the ~30
language-specific hook-template constants (`STUB_BOOTSTRAP_NODE`,
`STUB_LIFECYCLE_*` for V/Ruby/Bash/Bun/PHP/Lua/Deno/Go/Native,
`STUB_ASSEMBLE_*`, `STUB_COMPACT_*`, plus the long
`PY_INGEST`/`PY_AFTER_TURN`/`PY_ASSEMBLE`/`PY_COMPACT`/`PY_BOOTSTRAP`
and the Ruby / Bash / Bun / PHP / Lua `*_INGEST` and `*_AFTER_TURN`
templates).

The template constants alone were ~1,300 LOC of raw-string literals
embedded in the parent module — pure dead weight from the lint-noise
perspective, since each template is a single self-contained scaffold
artifact with no relationship to the surrounding install / verify /
publish logic.

scaffold.rs uses `use super::*;` to inherit `validate_plugin_name`,
`ensure_plugins_dir`, `plugins_dir`, `plugin_runtime::PluginRuntime`,
and the manifest types it consumes.

plugin_manager.rs: 4,725 -> 3,128 LOC (-1,597). Still over the 2,000
LOC cap from #3710's acceptance criteria; the next slice will pull out
the registry-verify / signed-index path (~600 LOC) and then the
install / publish / deps clusters.

cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
